pyproject.toml
[tool.poetry]
name = "nltk-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
sklearn = "^0.0"
keras = "^2.4.3"
en-core-web-sm-mirror = "^2.2.5"
spacy = "^3.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
